// frontend/src/util/enum/Enum.ts
export enum Role {
    GeneralUser = 0,
    Admin = 1,
    Engineer = 2
}

export enum RouterPath {
    Login = '/',
    Home = '/searchengineerlist',
    SkillSet = '/skillset',
    EngineerList = '/engineerlist',
    PasswordReset = '/passwordreset',
    CreateAccount = '/createaccount',
    MyPage = '/mypage'
}